Restyling: Forms and Buttons
This BEMifies forms and buttons. It replaces the automatic form from Django
with an explicit form, which is not necessarily the best thing to do, but
we will be rewriting the frontend to not be server-side and any potential
vulnerability will be short-lived. This does mean we need to update validation
though, but validation needs to be restyled still as well.